+++ CHANGELOG.md
@@ -5,6 +5,7 @@ v1.0.0 (2021.11.xx)
 -------------------
 - Lists lookup
   - Add Twitter.getList() for "GET /2/lists/:id"
+  - Add Twitter.getOwnedLists() for "GET /2/users/:id/owned_lists"
 - Bump Kotlin 1.6.0
 
 v0.4.0 (2021.11.12)